--- a/scripts/compare.py
+++ b/scripts/compare.py
@@ -1,13 +1,28 @@
 #!/usr/bin/env python3
 
 import sys
+import re
 
 logfile_1 = sys.argv[1]
 logfile_2 = sys.argv[2]
-mintime   = 0.01
-# minchange = 0.035
-minchange = 0.035
-  
+if len(sys.argv) > 3:
+    mintime = float(sys.argv[3])
+else:
+    mintime = 0.01
+if len(sys.argv) > 4:
+    minchange = float(sys.argv[4])
+else:
+    minchange = 0.035
+
+testname = re.compile(r'^src/.*\[(?P<name>.*)\]$')
+
+def readName(input):
+    r = testname.match(input)
+    if r is None:
+        return input
+    else:
+        return r.group("name")
+
 def readData(file_name):
     data_lines = []
     with open(file_name, 'r') as file_data:
@@ -15,7 +30,7 @@ def readData(file_name):
             data_line = [l for l in line.strip('\n').split(' ') if l != '']
             if data_line[1] == 'call':
                 time = float(data_line[0][:-1])
-                test = data_line[2]
+                test = readName(data_line[2])
                 if time > 0.5:
                     data_lines.append((test, time))
     return data_lines
@@ -35,7 +50,7 @@ def filterEntries(test, t1, t2):
     return (t1 != t2)                                                   \
        and (t1 > mintime or t2 > mintime)                               \
        and ((t1 / t2) - 1.0 > minchange or (t2 / t1) - 1.0 > minchange)
-    
+
 # test , time1 , time2 , time1 / time2
 final_table = []
 for test in common_passing_tests: