… test/rpc-integration/test-vacuous/response-vacuous-at-branch.json
… test/rpc-integration/test-vacuous/response-vacuous-but-rewritten.json
… test/rpc-integration/test-vacuous/response-vacuous-var-at-branch.json
… test/rpc-integration/test-vacuous/response-vacuous-without-rewrite.json