When an execute request ends with an `Aborted`, `Stuck`, or `Branching` reason, it is usually re-executed with legacy `kore-rpc`. As part of that, the simplifier will run on the input.

However, we are _also_ explicitly simplifying the end state of the execute request (and return `Vacuous` if it simplifies to `\bottom`). This additional work is unnecessary as we will anyway get the `\bottom` result back from `kore-rpc` (alas, in a slightly different shape).

This saves one round of fall-back to `kore-rpc` and should speed up long-running proofs that have to fall back often due to side condition checking.