@jberthold 2023-11-09
## KEVM comparison

| Test                                      | HOTFIX-avoid-kore-simplification time | master time       | (HOTFIX-avoid-kore-simplification/master) time 
|-------------------------------------------|---------------------------------------|-------------------|------------------------------------------------
| mcd/end-subuu-pass-spec.k                 | 67.29                                 | 143.81            | 0.4679090466587859                             
| erc20/ds/balanceOf-spec.k                 | 44.61                                 | 62.21             | 0.7170872850024111                             
| mcd/vat-subui-pass-rough-spec.k           | 159.03                                | 218.98            | 0.7262307060005481                             
| mcd/vat-addui-pass-spec.k                 | 159.24                                | 217.99            | 0.730492224413964                              
| mcd/vat-subui-fail-rough-spec.k           | 109.65                                | 148.49            | 0.7384335645498014                             
| mcd/vat-addui-fail-rough-spec.k           | 111.5                                 | 150.3             | 0.7418496340652029                             
| mcd/vat-mului-pass-spec.k                 | 123.46                                | 165.09            | 0.7478345145072384                             
| mcd/vat-muluu-pass-spec.k                 | 126.45                                | 167.32            | 0.7557375089648578                             
| mcd/vat-slip-pass-rough-spec.k            | 202.71                                | 264.78            | 0.7655789712213914                             
| benchmarks/address00-spec.k               | 43.48                                 | 56.07             | 0.7754592473693597                             
| benchmarks/ecrecover00-siginvalid-spec.k  | 78.93                                 | 101.3             | 0.7791707798617967                             
| mcd/vat-subui-pass-spec.k                 | 219.32                                | 279.31            | 0.7852207224947191                             
| mcd/flipper-bids-pass-rough-spec.k        | 132.51                                | 167.61            | 0.7905852872740289                             
| mcd/vat-move-diff-rough-spec.k            | 151.74                                | 184.36            | 0.8230635712735951                             
| bihu/forwardToHotWallet-success-1-spec.k  | 57.8                                  | 68.85             | 0.8395061728395062                             
| mcd/cat-exhaustiveness-spec.k             | 122.12                                | 143.53            | 0.8508325785550059                             
| mcd/dai-adduu-fail-rough-spec.k           | 60.15                                 | 70.59             | 0.8521036974075648                             
| mcd/flipper-addu48u48-fail-rough-spec.k   | 63.43                                 | 73.67             | 0.861001764626035                              
| bihu/forwardToHotWallet-failure-4-spec.k  | 94.46                                 | 108.48            | 0.8707595870206489                             
| mcd/dstoken-approve-fail-rough-spec.k     | 93.6                                  | 106.06            | 0.8825193286818781                             
| examples/solidity-code-spec.md            | 176.78                                | 198.93            | 0.8886543005077162                             
| bihu/forwardToHotWallet-failure-3-spec.k  | 74.87                                 | 83.91             | 0.8922655225837207                             
| bihu/forwardToHotWallet-failure-2-spec.k  | 61.99                                 | 69.29             | 0.8946456920190503                             
| mcd/dsvalue-peek-pass-rough-spec.k        | 59.84                                 | 66.47             | 0.9002557544757034                             
| mcd/dsvalue-read-pass-spec.k              | 58.26                                 | 64.43             | 0.9042371566040663                             
| erc20/hkg/transferFrom-failure-1-spec.k   | 75.82                                 | 82.87             | 0.9149269940871243                             
| erc20/ds/transfer-failure-1-b-spec.k      | 87.3                                  | 95.18             | 0.917209497793654                              
| bihu/forwardToHotWallet-success-2-spec.k  | 122.11                                | 133.12            | 0.9172926682692307                             
| mcd/flipper-ttl-pass-spec.k               | 57.37                                 | 62.49             | 0.9180668907025124                             
| erc20/ds/transfer-success-1-spec.k        | 67.2                                  | 73.01             | 0.9204218600191755                             
| erc20/ds/transferFrom-failure-2-a-spec.k  | 71.16                                 | 77.27             | 0.9209266209395626                             
| erc20/hkg/transferFrom-failure-2-spec.k   | 82.9                                  | 89.97             | 0.9214182505279538                             
| mcd/flopper-cage-pass-spec.k              | 59.47                                 | 64.38             | 0.9237340789064927                             
| erc20/ds/transferFrom-success-2-spec.k    | 80.72                                 | 87.33             | 0.9243100881713042                             
| erc20/ds/transferFrom-failure-1-b-spec.k  | 128.56                                | 138.9             | 0.9255579553635709                             
| mcd/vat-deny-diff-fail-rough-spec.k       | 88.36                                 | 95.24             | 0.9277614447711047                             
| erc20/hkg/transfer-success-2-spec.k       | 51.44                                 | 55.38             | 0.928855182376309                              
| erc20/hkg/transferFrom-success-1-spec.k   | 68.42                                 | 73.52             | 0.9306311207834603                             
| erc20/ds/transferFrom-failure-1-c-spec.k  | 98.08                                 | 105.29            | 0.931522461772248                              
| erc20/ds/transfer-failure-1-a-spec.k      | 67.53                                 | 72.47             | 0.9318338622878433                             
| erc20/ds/transferFrom-failure-1-a-spec.k  | 82.39                                 | 88.32             | 0.9328577898550725                             
| mcd/flipper-tau-pass-spec.k               | 58.07                                 | 62.22             | 0.9333011893281903                             
| erc20/ds/transfer-success-2-spec.k        | 63.92                                 | 68.44             | 0.9339567504383403                             
| erc20/ds/transferFrom-failure-2-b-spec.k  | 86.78                                 | 92.63             | 0.9368455144121776                             
| erc20/ds/transferFrom-success-1-spec.k    | 92.93                                 | 99.15             | 0.9372667675239537                             
| benchmarks/storagevar02-overflow-spec.k   | 43.28                                 | 46.03             | 0.9402563545513796                             
| erc20/hkg/transfer-failure-2-spec.k       | 53.2                                  | 56.47             | 0.9420931468036126                             
| bihu/forwardToHotWallet-failure-1-spec.k  | 45.09                                 | 47.69             | 0.9454812329628854                             
| erc20/hkg/transfer-success-1-spec.k       | 54.42                                 | 57.35             | 0.9489102005231037                             
| mcd/cat-file-addr-pass-rough-spec.k       | 57.77                                 | 60.78             | 0.9504771306350773                             
| benchmarks/storagevar02-nooverflow-spec.k | 44.52                                 | 46.69             | 0.9535232383808097                             
| examples/storage-spec.md                  | 163.7                                 | 171.65            | 0.9536848237692979                             
| functional/evm-int-simplifications-spec.k | 97.02                                 | 101.7             | 0.9539823008849557                             
| benchmarks/encode-keccak00-spec.k         | 49.9                                  | 52.23             | 0.9553896228221329                             
| mcd/dai-symbol-pass-spec.k                | 50.14                                 | 52.38             | 0.9572355861015654                             
| erc20/ds/transfer-failure-2-a-spec.k      | 66.52                                 | 69.27             | 0.960300274289014                              
| erc20/hkg/allowance-spec.k                | 42.98                                 | 44.73             | 0.9608763693270735                             
| mcd/functional-spec.k                     | 417.99                                | 434.77            | 0.9614048807415415                             
| erc20/ds/approve-failure-spec.k           | 46.28                                 | 48.01             | 0.9639658404499063                             
| erc20/ds/transferFrom-failure-2-c-spec.k  | 53.67                                 | 55.61             | 0.9651141880956663                             
| benchmarks/storagevar01-spec.k            | 59.28                                 | 61.38             | 0.9657869012707723                             
| erc20/hkg/transfer-failure-1-spec.k       | 55.46                                 | 57.41             | 0.9660337920222959                             
| examples/erc20-spec.md                    | 206.9                                 | 193.76            | 1.0678158546655658                             
| benchmarks/ecrecover00-sigvalid-spec.k    | 101.56                                | 95.11             | 1.0678162128062243                             
| erc20/ds/transfer-failure-2-b-spec.k      | 62.64                                 | 49.64             | 1.2618855761482675                             
| benchmarks/bytes00-spec.k                 | 69.43                                 | 54.77             | 1.2676647799890453                             
| mcd/dsvalue-read-pass-summarize-spec.k    | 111.72                                | 59.09             | 1.8906752411575563                             
| TOTAL                                     | 6197.220000000002                     | 7015.529999999997 | 0.8833573514759405